airflow/requirements.txt
httpx
tenacity
psycopg2-binary
redis
